исправил ошибку в CodeStyle,в InMemoryUserStorage 2

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.UserNotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
@@ -12,11 +13,13 @@ import java.util.Collection;
 public class FilmService {
 
     private final FilmStorage filmStorage;
+    private final UserStorage userStorage;
 
 
     @Autowired
     public FilmService(FilmStorage filmStorage, UserStorage userStorage) {
         this.filmStorage = filmStorage;
+        this.userStorage =userStorage;
     }
 
     public Collection<Film> findAllFilms() {
@@ -32,10 +35,16 @@ public class FilmService {
     }
 
     public Film addLike(int id, int userId) {
+        if (!userStorage.checkUser(userId)) {
+            throw new UserNotFoundException("Пользователь с id=" + userId + " не найден");
+        }
         return filmStorage.addLike(id, userId);
     }
 
     public Film deleteLike(int id, int userId) {
+        if (!userStorage.checkUser(userId)) {
+            throw new UserNotFoundException("Пользователь с id=" + userId + " не найден");
+        }
         return filmStorage.deleteLike(id, userId);
     }
